requirements.txt
# Flask web framework for Azure Web App
Flask>=2.3.0

# Azure Functions (if still needed)
azure-functions>=1.18.0

# Azure Cosmos DB client
azure-cosmos>=4.5.0

# Core Python libraries (usually included but good to specify)
uuid

# Optional: For better JSON handling and debugging
requests>=2.31.0
